refactor(notes): Retire pads session manager
