+++ bigbluebutton-html5/imports/ui/components/pads/pads-graphql/component.tsx
@@ -1,8 +1,12 @@
 import React, { useEffect, useState } from 'react';
 import { defineMessages, useIntl } from 'react-intl';
 import { useMutation } from '@apollo/client';
-import { HAS_PAD_SUBSCRIPTION, HasPadSubscriptionResponse } from './queries';
-import { PAD_SESSION_SUBSCRIPTION, PadSessionSubscriptionResponse } from './sessions/queries';
+import {
+  HAS_PAD_SUBSCRIPTION,
+  HasPadSubscriptionResponse,
+  PAD_SESSION_SUBSCRIPTION,
+  PadSessionSubscriptionResponse,
+} from './queries';
 import { CREATE_SESSION } from './mutations';
 import Service from './service';
 import Styled from './styles';
