+++ bigbluebutton-html5/imports/ui/components/pads/pads-graphql/queries.ts
@@ -16,6 +16,31 @@ export const HAS_PAD_SUBSCRIPTION = gql`
   }
 `;
 
+export interface PadSessionSubscriptionResponse {
+  sharedNotes_session: Array<{
+    sessionId: string;
+    sharedNotesExtId: string;
+    padId: string;
+    sharedNotes: {
+      padId: string;
+    };
+  }>;
+}
+
+export const PAD_SESSION_SUBSCRIPTION = gql`
+  subscription padSession {
+    sharedNotes_session {
+      sessionId
+      sharedNotesExtId
+      padId
+      sharedNotes {
+        padId
+      }
+    }
+  }
+`;
+
 export default {
   HAS_PAD_SUBSCRIPTION,
+  PAD_SESSION_SUBSCRIPTION,
 };
