--- bigbluebutton-html5/imports/ui/components/pads/pads-graphql/sessions/service.ts
@@ -1,13 +1,0 @@
-const setCookie = (sessions: Array<string>) => {
-  const COOKIE_CONFIG = window.meetingClientSettings.public.pads.cookie;
-  const PATH = COOKIE_CONFIG.path;
-  const SAME_SITE = COOKIE_CONFIG.sameSite;
-  const SECURE = COOKIE_CONFIG.secure;
-
-  const sessionIds = sessions.join(',');
-  document.cookie = `sessionID=${sessionIds}; path=${PATH}; SameSite=${SAME_SITE}; ${SECURE ? 'Secure' : ''}`;
-};
-
-export default {
-  setCookie,
-};
