+++ bigbluebutton-html5/imports/ui/core/initial-values/meetingClientSettings.ts
@@ -613,11 +613,6 @@ export const meetingClientSettingsInitialValues: MeetingClientSettings = {
     },
     pads: {
       url: 'ETHERPAD_HOST',
-      cookie: {
-        path: '/',
-        sameSite: 'None',
-        secure: true,
-      },
     },
     media: {
       audio: {
